Began attempt to prepare prototype
I began by adding some comments and trying to clean up the code. I then realized there is a good bit to be cleaned up. I started to look into refactoring some code with some base class interfaces, but the amount of things needed to be refactored grew and grew. So For now this is a saving spot incase I try again, but I might revert and leave this technical debt to another day.