Fixed UML Classes Diagram #121
There was one place in my last commit where I removed a stray import ProfileController which wasnt at the top of the document. This caused the one at the top to be highlighted as if it wasn't used. Removing the secondary import later in the document hasn't lead to any errors and there were no comments explaining the strange placement of the import. But it also corrected me, because I thought that ProfileController wasn't being used because of the highlighted import. Anyway I updated the UML to correct my mistake.