Fixed mismatched Behavior
To do this I made the buttons implement the same function calls to the model and the CombinedDataController that already existed on the mainController Object. There were some redundant checks for Nones that also were gotten rid of. I changed the return values in the combine_data and load_combined_data functions to return a success boolean. I was sure to track all the references to those functions and ensured there was no side effects of this change. Because there was only one reference in the main controller it worked out nicely.  There was also an occational error when we check for duplicate data thats not there because we werent initializing duplicate_data in the data_model even though we were checking it to be none. So I went ahead and added that to the init as well.