Styles data class made for Views
Created view_constants.py which holds constants like window width and height as well as various colors for the app. Then I applied this class in all of the instances in the views. Now if we want to make changes to the colors or sizes we only need to change the value in the view_constants file inorder to update the app.