Made it so Init view is hidden after loads combined data
just needed to add a internal reference to it's view so It could call it on remove tab. This is a quickfix I know we considered the idea of having multiple combined views operational at once, but until we have time to explore that further this will keep users from causing unintentional behavior.