Updates to my docs, slight changes to files to clean up, and folder renames
Made a Milestone 4 folder and renamed the milestone1 and milestone2 to have a space in their name so they are all named similarly. Then I copied the swot analysis and then started updating the UML diagram. While updating the UML I made a few clean ups to either make the code align with the uml or just cleaning unusing functions or comments.